ToolServer/requirements.txt
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
django-cors-headers
debugpy==1.6.7
decorator==5.1.1
django-cors-headers==3.14.0
djangorestframework==3.14.0
django==5.0
exceptiongroup==1.1.1
executing==1.2.0
idna==3.4
ipython==8.14.0
jedi==0.19.0
jwt==1.3.0
matplotlib-inline==0.1.6
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.8.1
prompt-toolkit==3.0.39
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.1.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tornado==6.3.2
traitlets==5.9.0
typing-extensions==4.7.1
urllib3==2.0.3
wcwidth==0.2.6